+++ src/components/team-users/styles.styl
@@ -1,3 +1,5 @@
+@import '../constants.styl'
+
 .teamUserWrap
   margin-right: -7px
   &.lastTeamUser
@@ -11,29 +13,6 @@
 .addUser
   transition: all .2s
   
-.whitelistedDomainContainer
-  margin: 5px 5px 0
-  
-.whitelistedDomainIcon,
-.whitelistedDomainLabel
-  width: 22px
-  height: 22px
-  
-.whitelistedDomainIcon
-  display: inline-block
-  object-fit: contain
-  margin: auto
-      
-.whitelistedDomainLabel
-  display: inline-flex
-  justify-content: center
-  align-items: center
-  background-color: line-on-secondary-background
-  font-size: 13px
-  font-weight: 600
-  color: white
-  border-radius: 3px
-  
 .container
   margin: 0
   padding: 0
@@ -41,4 +20,3 @@
   display: flex
   flex-flow: row wrap
   align-items: center
-
