+++ src/components/whitelisted-domain/styles.styl
@@ -1,0 +1,24 @@
+@import '../constants.styl'
+
+.whitelistedDomainContainer
+  margin: 5px 5px 0
+
+.whitelistedDomainIcon,
+.whitelistedDomainLabel
+  width: 22px
+  height: 22px
+  
+.whitelistedDomainIcon
+  display: inline-block
+  object-fit: contain
+  margin: auto
+      
+.whitelistedDomainLabel
+  display: inline-flex
+  justify-content: center
+  align-items: center
+  background-color: line-on-secondary-background
+  font-size: 13px
+  font-weight: 600
+  color: white
+  border-radius: 3px
